Extract pure table geometry helpers into view/table-geometry
Move pure-geometry helpers out of table-renderer.ts (and a copy of
computeTableRangeForPageLine out of doc-canvas.ts) into a new
view/table-geometry.ts module so the upcoming PDF table painter can
share them with the Canvas renderer.

No behavior change: table-renderer.ts re-exports the moved symbols
for existing call sites, table-layout.ts now imports
computeMergedCellLineLayouts directly from table-geometry, and
doc-canvas.ts uses the shared computeTableRangeForPageLine. All 661
docs tests still pass.
